+++ src/Adapter/Place/PlaceWithPostalCode.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./Place.mjs").Place} Place */
+
+/**
+ * @typedef {Place & {"postal-code": number}} PlaceWithPostalCode
+ */