+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-01-03-3
+
+Changes:
+
+- Show identification number
+
 ## v2023-01-03-2
 
 Changes: